Fixed external dependency management (hopefully).
Corrected the repository layout for new external dependencies. More
importantly, adapted the file URL to the external-jars directory in a
way that allows each module to be built independently, I hope. The
formerly given form "file://..." seems to have references an absolute
path (not sure) even when a '.' or a '..' was given as a first element,
which should introduce a relative path
(https://tools.ietf.org/html/rfc3986#section-4.2). Or I didn't
understand that right. Well, this way it seems to work.